+++ lib/article_json.rb
@@ -49,6 +49,7 @@ require_relative 'article_json/export/common/html/elements/list'
 require_relative 'article_json/export/common/html/elements/image'
 require_relative 'article_json/export/common/html/elements/quote'
 require_relative 'article_json/export/common/html/elements/text_box'
+require_relative 'article_json/export/common/html/elements/embed'
 
 require_relative 'article_json/export/html/elements/base'
 require_relative 'article_json/export/html/elements/text'
