+++ lib/article_json/export/html/elements/embed.rb
@@ -3,28 +3,7 @@ module ArticleJSON
     module HTML
       module Elements
         class Embed < Base
-          include ArticleJSON::Export::Common::HTML::Elements::Shared::Caption
-
-          # Generate the embedded element node
-          # @return [Nokogiri::XML::NodeSet]
-          def export
-            create_element(:figure) do |figure|
-              figure.add_child(embed_node)
-              figure.add_child(caption_node(:figcaption))
-            end
-          end
-
-          private
-
-          def embed_node
-            create_element(:div, class: 'embed') do |div|
-              div.add_child(embedded_object)
-            end
-          end
-
-          def embedded_object
-            Nokogiri::HTML.fragment(@element.oembed_data[:html])
-          end
+          include ArticleJSON::Export::Common::HTML::Elements::Embed
         end
       end
     end
